Ensure that a valid JSONPath never fails to execute (#27)
In case no value is matched by a JSONPath or it doesn't match the structure of
the JSON, it is no longer an "error". This arguably makes the library lose a
"feature", but having JSONPath result in errors causes problems when there is
trailing path to follow or when the path is used in a search (after `..`) or
filter (inside `[?(...)]`) context .

This change also gets rid of the `ExecutionResult` type, as it was a little
ambiguous when a list should be returned and when a scalar value should be
returned. Most libraries in other languages which return scalar values when
possible, also return `null` when the value doesn't match. If
`executeJSONPathElement` were to return a `null` for these, the search and
filter operations need to have special handling for `null` as they should now
filter them out. But even that is ambiguous, because there could be `null` value
in an array and it must not be filtered out. So, in order to make things
simpler, it is best to assume that execution of a JSONPath will return a list.
This list will be empty when there are no matches, if the list contains nulls,
it just means that the JSONPath matched some null values in the provided JSON.