Parser: Explicitly look for ending condition to get better errors
Without this, if any path element after the first one is malformed, the jsonPath
parser stops without consuming the malformed element. If no parser follows this
half parsed expression is returned if there is some parser afterwards (say eof),
the error only mentions that eof is expected and not anything about a json path
element.